requirements.txt
PyYAML>=5.3
orjson>=3,<4